+++ sdk/ai/azopenai/CHANGELOG.md
@@ -4,9 +4,15 @@
 
 ### Features Added
 
+- `StreamOptions` parameter added to `ChatCompletionsOptions` and `CompletionsOptions`.
+- `MaxCompletionTokens` parameter added to `ChatCompletionsOptions`.
+- `ParallelToolCalls` parameter added to `ChatCompletionsOptions`.
+
 ### Breaking Changes
 
 - `MongoDBChatExtensionParameters.Authentication`'s type has been changed to a `OnYourDataUsernameAndPasswordAuthenticationOptions`. (PR#23620)
+- `GetCompletions` and `GetCompletionsStream` now receive different options (`CompletionsOptions` and `CompletionsStreamOptions` respectively)
+- `GetChatCompletions` and `GetChatCompletionsStream` now receive different options (`ChatCompletionsOptions` and `ChatCompletionsStreamOptions` respectively)
 
 ### Bugs Fixed