+++ sdk/resourcemanager/springappdiscovery/armspringappdiscovery/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release History
 
 ## 0.1.0 (2024-02-01)
+### Other Changes
 
 The package of `github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/springappdiscovery/armspringappdiscovery` is using our [next generation design principles](https://azure.github.io/azure-sdk/general_introduction.html).